+++ hil/hil.hh
@@ -91,7 +91,7 @@ class HIL : public Object {
   LPN getTotalPages();
 
   //! Get bytesize of one logical page.
-  uint64_t getLPNSize();
+  uint32_t getLPNSize();
 
   //! Enable/disable ICL
   void setCache(bool);
